1.解决场景移除全部玩家没有移除完问题，2.解决ushort到uint的序列化问题...
1.解决场景移除全部玩家没有移除完问题，2.解决ushort到uint的序列化问题 3.优化缓存类和查询功能内置 4.分布式案例更新

+++ GameDesigner/Network/core/Client/ClientBase.cs
@@ -1936,8 +1936,8 @@ namespace Net.Client
         #region 远程过程调用
         public void Call(uint protocol, params object[] pars) => Call(NetCmd.CallRpc, protocol, null, pars);
         public void Call(byte cmd, uint protocol, params object[] pars) => Call(cmd, protocol, null, pars);
-        public void Call(byte[] buffer) => Call(NetCmd.OtherCmd, (ushort)0, buffer);
-        public void Call(byte cmd, byte[] buffer) => Call(cmd, (ushort)0, buffer);
+        public void Call(byte[] buffer) => Call(NetCmd.OtherCmd, 0U, buffer);
+        public void Call(byte cmd, byte[] buffer) => Call(cmd, 0U, buffer);
         public void Call(string func, params object[] pars) => Call(NetCmd.CallRpc, func.CRCU32(), null, pars);
         public void Call(byte cmd, string func, params object[] pars) => Call(cmd, func.CRCU32(), null, pars);
