+++ GameDesigner/Network/core/Serialize/NetConvertFast.cs
@@ -189,7 +189,7 @@
             FuncData fdata = default;
             try
             {
-                fdata.protocol = segment.ReadUInt16();
+                fdata.protocol = segment.ReadUInt32();
                 var list = new List<object>();
                 while (segment.Position < segment.Offset + segment.Count)
                 {
